Authentication and Authorization

+++ frontend/src/components/CourseDetails.js
@@ -1,12 +1,17 @@
 import { useCoursesContext } from "../hooks/useCourseContext"
+import { useAuthContext } from "../hooks/useAuthContext"
 
 // date fns
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 
 const CourseDetails = ({ course }) => {
   const { dispatch } = useCoursesContext()
+  const {user} = useAuthContext()
 
   const handleClick = async () => {
+    if(!user){
+      return
+    }
     const response = await fetch('/course/' + course._id, {
       method: 'DELETE'
     })
